bus/pci: remove unnecessary cast
The variables 'vfio_res->nb_maps' and 'i' are of type int. The type
casting of 'vfio_res->nb_maps' is redundant and not required.

Acked-by: Anatoly Burakov <anatoly.burakov@intel.com>

+++ drivers/bus/pci/linux/pci_vfio.c
@@ -762,7 +762,7 @@ pci_vfio_map_resource_primary(struct rte_pci_device *dev)
 		}
 	}
 
-	for (i = 0; i < (int) vfio_res->nb_maps; i++) {
+	for (i = 0; i < vfio_res->nb_maps; i++) {
 		struct vfio_region_info *reg = NULL;
 		void *bar_addr;
 
@@ -888,7 +888,7 @@ pci_vfio_map_resource_secondary(struct rte_pci_device *dev)
 	/* map BARs */
 	maps = vfio_res->maps;
 
-	for (i = 0; i < (int) vfio_res->nb_maps; i++) {
+	for (i = 0; i < vfio_res->nb_maps; i++) {
 		ret = pci_vfio_mmap_bar(vfio_dev_fd, vfio_res, i, MAP_FIXED);
 		if (ret < 0) {
 			RTE_LOG(ERR, EAL, "  %s mapping BAR%i failed: %s\n",
@@ -948,7 +948,7 @@ find_and_unmap_vfio_resource(struct mapped_pci_res_list *vfio_res_list,
 		pci_addr);
 
 	maps = vfio_res->maps;
-	for (i = 0; i < (int) vfio_res->nb_maps; i++) {
+	for (i = 0; i < vfio_res->nb_maps; i++) {
 
 		/*
 		 * We do not need to be aware of MSI-X table BAR mappings as
